Sesiones y aulas

+++ sitio/cake/app/Controller/UsuariosController.php
@@ -6,6 +6,7 @@ App::uses('AppController', 'Controller');
  * @property Usuario $Usuario
  * @property PaginatorComponent $Paginator
  */
+ session_start();
 class UsuariosController extends AppController {
 
 /**
@@ -23,8 +24,16 @@ class UsuariosController extends AppController {
  * @return void
  */
 	public function index() {
-		$this->Usuario->recursive = 0;
-		$this->set('usuarios', $this->Paginator->paginate());
+        if(isset($_SESSION['tipo_usuario']) and $_SESSION['tipo_usuario']<=2 )
+        {
+		    $this->Usuario->recursive = 0;
+		    $this->set('usuarios', $this->Paginator->paginate());
+        }
+        else
+        {
+                $this->redirect(array('controller' =>'inicio','action' => 'index'));    
+        }
+
 	}
 
 /**
@@ -35,11 +44,18 @@ class UsuariosController extends AppController {
  * @return void
  */
 	public function view($id = null) {
-		if (!$this->Usuario->exists($id)) {
-			throw new NotFoundException(__('Usuario Inválido'));
-		}
-		$options = array('conditions' => array('Usuario.' . $this->Usuario->primaryKey => $id));
-		$this->set('usuario', $this->Usuario->find('first', $options));
+        if(isset($_SESSION['tipo_usuario']) and $_SESSION['tipo_usuario']<=2 )
+        {
+		    if (!$this->Usuario->exists($id)) {
+			    throw new NotFoundException(__('Usuario Inválido'));
+		    }
+		    $options = array('conditions' => array('Usuario.' . $this->Usuario->primaryKey => $id));
+		    $this->set('usuario', $this->Usuario->find('first', $options));
+        }
+        else
+        {
+                $this->redirect(array('controller' =>'inicio','action' => 'index'));    
+        }
 	}
 
 /**
@@ -49,45 +65,58 @@ class UsuariosController extends AppController {
  */
 	public function add() {		
 		$temp;
-		
-		if ($this->request->is('post')) {
+		if(isset($_SESSION['tipo_usuario']) and $_SESSION['tipo_usuario']<=1 )
+        {
+		    if ($this->request->is('post')) {
 			
-			if ($this->request->data['Usuario']['nick'] == "") {
-				$nick = $this->crearNick($this->request->data['Usuario']['nombre'],
-										 $this->request->data['Usuario']['apellidos']);
-				$this->request->data['Usuario']['nick'] = $nick;
-			}
+			    if ($this->request->data['Usuario']['nick'] == "") {
+				    $nick = $this->crearNick($this->request->data['Usuario']['nombre'],
+										     $this->request->data['Usuario']['apellidos']);
+				    $this->request->data['Usuario']['nick'] = $nick;
+			    }
 			
-			if ($this->request->data['Usuario']['contrasena'] == "") {
-				$this->request->data['Usuario']['contrasena'] = $this->crearContrasena($this->request->data['Usuario']['nombre'],
-																					   $this->request->data['Usuario']['apellidos']);	
-			}
+			    if ($this->request->data['Usuario']['contrasena'] == "") {
+				    $this->request->data['Usuario']['contrasena'] = $this->crearContrasena($this->request->data['Usuario']['nombre'],
+																					       $this->request->data['Usuario']['apellidos']);	
+			    }
 			
-			//para el md5 de la contrasena de tamano 32
-			$this->request->data['Usuario']['contrasena'] = md5($this->request->data['Usuario']['contrasena']);
-			$this->request->data['Usuario']['fecha'] = date("Y-m-d H:i:s");
+			    //para el md5 de la contrasena de tamano 32
+			    $this->request->data['Usuario']['contrasena'] = md5($this->request->data['Usuario']['contrasena']);
+			    $this->request->data['Usuario']['fecha'] = date("Y-m-d H:i:s");
 			
-			//creacion del usuario para ser almacenado
-			$this->Usuario->create();
-			if ($this->Usuario->save($this->request->data)) {
-				$this->Session->setFlash(__('El usuario ha sido creado.'));
-				return $this->redirect(array('action' => 'index'));
-			} else {
-				$this->Session->setFlash(__('El usuario no se ha podido crear. Por favor, intente de nuevo.'));
-			}
-		}
+			    //creacion del usuario para ser almacenado
+			    $this->Usuario->create();
+			    if ($this->Usuario->save($this->request->data)) {
+				    $this->Session->setFlash(__('El usuario ha sido creado.'));
+				    return $this->redirect(array('action' => 'index'));
+			    } else {
+				    $this->Session->setFlash(__('El usuario no se ha podido crear. Por favor, intente de nuevo.'));
+			    }
+		    }
+        }
+        else
+        {
+                $this->redirect(array('controller' =>'inicio','action' => 'index'));    
+        }
 	}
 	
 	public function addContinue() {
-		if ($this->request->is('post')) {
-			$this->Usuario->create();
-			if ($this->Usuario->save($this->request->data)) {
-				$this->Session->setFlash(__('El usuario ha sido guardado.'));
-				return $this->redirect(array('action' => 'add'));
-			} else {
-				$this->Session->setFlash(__('El usuario no se ha podido guardar. Por favor, intente de nuevo.'));
-			}
-		}
+        if(isset($_SESSION['tipo_usuario']) and $_SESSION['tipo_usuario']<=1 )
+        {
+		    if ($this->request->is('post')) {
+			    $this->Usuario->create();
+			    if ($this->Usuario->save($this->request->data)) {
+				    $this->Session->setFlash(__('El usuario ha sido guardado.'));
+				    return $this->redirect(array('action' => 'add'));
+			    } else {
+				    $this->Session->setFlash(__('El usuario no se ha podido guardar. Por favor, intente de nuevo.'));
+			    }
+		    }
+        }
+        else
+        {
+                $this->redirect(array('controller' =>'inicio','action' => 'index'));    
+        }
 	}
 
 /**
@@ -99,40 +128,46 @@ class UsuariosController extends AppController {
  */
 	public function edit($id = null) {
 		$temp;
-		
-		if (!$this->Usuario->exists($id)) {
-			throw new NotFoundException(__('Usuario Inválido'));
-		}
-		if ($this->request->is(array('post', 'put'))) {
-			if ($this->request->data['Usuario']['contrasena'] == $this->request->data['Usuario']['confirmarContrasena']) {
-				if ($this->request->data['Usuario']['nick'] == "") {
-					$nick = $this->crearNick($this->request->data['Usuario']['nombre'],
-											 $this->request->data['Usuario']['apellidos']);
-					$this->request->data['Usuario']['nick'] = $nick;
-				}
+		if(isset($_SESSION['tipo_usuario']) and $_SESSION['tipo_usuario']<=1 )
+        {
+		    if (!$this->Usuario->exists($id)) {
+			    throw new NotFoundException(__('Usuario Inválido'));
+		    }
+		    if ($this->request->is(array('post', 'put'))) {
+			    if ($this->request->data['Usuario']['contrasena'] == $this->request->data['Usuario']['confirmarContrasena']) {
+				    if ($this->request->data['Usuario']['nick'] == "") {
+					    $nick = $this->crearNick($this->request->data['Usuario']['nombre'],
+											     $this->request->data['Usuario']['apellidos']);
+					    $this->request->data['Usuario']['nick'] = $nick;
+				    }
 				
-				if ($this->request->data['Usuario']['contrasena'] == "") {
-					$this->request->data['Usuario']['contrasena'] = $this->crearContrasena($this->request->data['Usuario']['nombre'],
-																						   $this->request->data['Usuario']['apellidos']);	
-				}
+				    if ($this->request->data['Usuario']['contrasena'] == "") {
+					    $this->request->data['Usuario']['contrasena'] = $this->crearContrasena($this->request->data['Usuario']['nombre'],
+																						       $this->request->data['Usuario']['apellidos']);	
+				    }
 				
-				//para el md5 de la contrasena de tamano 32
-				$this->request->data['Usuario']['contrasena'] = md5($this->request->data['Usuario']['contrasena']);
-				$this->request->data['Usuario']['fecha'] = date("Y-m-d H:i:s");
+				    //para el md5 de la contrasena de tamano 32
+				    $this->request->data['Usuario']['contrasena'] = md5($this->request->data['Usuario']['contrasena']);
+				    $this->request->data['Usuario']['fecha'] = date("Y-m-d H:i:s");
 				
-				if ($this->Usuario->save($this->request->data)) {
-					$this->Session->setFlash(__('El usuario ha sido actualizado'));
-					return $this->redirect(array('action' => 'index'));
-				} else {
-					$this->Session->setFlash(__('El usuario no se ha podido guardar. Por favor, intente de nuevo.'));
-				}
-			} else {
-				$this->Session->setFlash(__('Las contraseñas no coinciden.'));
-			}
-		} else {
-			$options = array('conditions' => array('Usuario.' . $this->Usuario->primaryKey => $id));
-			$this->request->data = $this->Usuario->find('first', $options);
-		}
+				    if ($this->Usuario->save($this->request->data)) {
+					    $this->Session->setFlash(__('El usuario ha sido actualizado'));
+					    return $this->redirect(array('action' => 'index'));
+				    } else {
+					    $this->Session->setFlash(__('El usuario no se ha podido guardar. Por favor, intente de nuevo.'));
+				    }
+			    } else {
+				    $this->Session->setFlash(__('Las contraseñas no coinciden.'));
+			    }
+		    } else {
+			    $options = array('conditions' => array('Usuario.' . $this->Usuario->primaryKey => $id));
+			    $this->request->data = $this->Usuario->find('first', $options);
+		    }
+        }
+        else
+        {
+                $this->redirect(array('controller' =>'inicio','action' => 'index'));    
+        }
 	}
 
 /**
@@ -143,31 +178,44 @@ class UsuariosController extends AppController {
  * @return void
  */
 	public function delete($id = null) {
-		$this->Usuario->id = $id;
-		if (!$this->Usuario->exists()) {
-			throw new NotFoundException(__('Invalid usuario'));
-		}
-		$this->request->allowMethod('post', 'delete');
-		if ($this->Usuario->delete()) {
-			$this->Session->setFlash(__('El usuario ha sido eliminado.'));
-		} else {
-			$this->Session->setFlash(__('El usuario no se ha podido eliminar. Por favor, intente de nuevo.'));
-		}
-		return $this->redirect(array('action' => 'index'));
+        if(isset($_SESSION['tipo_usuario']) and $_SESSION['tipo_usuario']<=1 )
+        {
+		    $this->Usuario->id = $id;
+		    if (!$this->Usuario->exists()) {
+			    throw new NotFoundException(__('Invalid usuario'));
+		    }
+		    $this->request->allowMethod('post', 'delete');
+		    if ($this->Usuario->delete()) {
+			    $this->Session->setFlash(__('El usuario ha sido eliminado.'));
+		    } else {
+			    $this->Session->setFlash(__('El usuario no se ha podido eliminar. Por favor, intente de nuevo.'));
+		    }
+		    return $this->redirect(array('action' => 'index'));
+        }
+        else
+        {
+             $this->redirect(array('controller' =>'inicio','action' => 'index'));    
+        }
 	}
 	
 	public function AjaxAddUser() {
 		$this->layout = 'ajax';
-		
-		if($this->request->is('post'))	{
-			$this->request->data['Usuario']['fecha'] = date("Y-m-d H:i:s");
+		if(isset($_SESSION['tipo_usuario']) and $_SESSION['tipo_usuario']<=1 )
+        {
+		    if($this->request->is('post'))	{
+			    $this->request->data['Usuario']['fecha'] = date("Y-m-d H:i:s");
 				
-			$this->Blogforo->Comentario->create();
+			    $this->Blogforo->Comentario->create();
 			
-			if($this->Blogforo->Comentario->save($this->request->data))	{
-				//echo ($this->request->data['Comentario']['comentario']);
-			}
-		}
+			    if($this->Blogforo->Comentario->save($this->request->data))	{
+				    //echo ($this->request->data['Comentario']['comentario']);
+			    }
+		    }
+        }
+        else
+        {
+            $this->redirect(array('controller' =>'inicio','action' => 'index'));    
+        }
 	}
 	
 	//para crear el nombre de usuario con el nombre y apellidos.
